git-gui: Perform rescan on window focus-in
If any changes are made to the tree while git-gui is open, the user has
to manually rescan to see those changes in the gui. With this change, a
rescan will be performed whenever the window comes in focus, removing
the need for manual rescans in most cases. A manual rescan will still be
needed when something makes changes to the tree while git-gui is still
in focus.

Since a rescan can potentially be a very expensive operation, it is
disabled by default, and can be enabled in the settings.

Signed-off-by: Pratyush Yadav <me@yadavpratyush.com>